fix: compare winId consistently in _q_onWindowMotifHintsChanged
The slot receives winId as quint32 (eventFilter and the DWindowManagerHelper
windowMotifWMHintsChanged signal both deliver it truncated to 32 bits), but the
identity guard compared it against the full-width WId from QWindow::winId().
On X11 WId is a 32-bit XID so it worked; on Wayland/Qt6.11 WId is a 64-bit
pointer (>4GiB), so the guard is always true and the slot returns before
fetching Motif functions, leaving motifFunctions at 0 and disabling the
minimize/close buttons on DTK client titlebars (e.g. dde-control-center).

+++ src/dquickwindow.cpp
@@ -214,7 +214,7 @@ void DQuickWindowAttachedPrivate::_q_onWindowMotifHintsChanged(quint32 winId)
 
     if (!q->window())
         return;
-    if (q->window()->winId() != winId)
+    if (static_cast<quint32>(q->window()->winId()) != winId)
         return;
 
     auto functions_hints = DWindowManagerHelper::getMotifFunctions(q->window());
